requirements.txt
absl-py==0.10.0
aio-pika==6.8.2
aiofiles==23.2.1
aiohttp==3.6.3
aiormq==3.3.1
APScheduler==3.6.3
astunparse==1.6.3
async-generator==1.10
async-timeout==3.0.1
attrs==20.2.0
backports.zoneinfo==0.2.1
boto3==1.33.6
botocore==1.33.6
cachetools==5.3.2
certifi @ file:///C:/b/abs_85o_6fm0se/croot/certifi_1671487778835/work/certifi
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==1.4.1
colorama==0.4.6
colorclass==2.2.2
coloredlogs==14.3
colorhash==1.0.4
cryptography==41.0.7
cycler==0.11.0
decorator==4.4.2
dm-tree==0.1.8
dnspython==1.16.0
docopt==0.6.2
fbmessenger==6.0.0
filelock==3.12.2
fsspec==2023.1.0
future==0.18.3
gast==0.3.3
google-auth==2.24.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.59.3
h11==0.9.0
h5py==2.10.0
httpcore==0.11.1
httplib2==0.22.0
httptools==0.6.0
httpx==0.15.4
huggingface-hub==0.16.4
humanfriendly==10.0
idna==3.6
importlib-metadata==6.7.0
jmespath==1.0.1
joblib==0.15.1
jsonpickle==1.4.2
jsonschema==3.2.0
kafka-python==2.0.2
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.3.4
mattermostwrapper==2.2
multidict==4.7.6
networkx==2.5.1
nltk==3.8.1
numpy==1.18.5
oauth2client==4.1.3
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==20.9
pamqp==2.3.0
Pillow==9.5.0
prompt-toolkit==2.0.10
protobuf==3.20.3
psycopg2-binary==2.8.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydot==1.4.2
PyJWT==1.7.1
pykwalify==1.7.0
pymongo==3.10.1
pyparsing==3.1.1
pyreadline==2.1
pyrsistent==0.19.3
python-crfsuite==0.9.9
python-dateutil==2.8.2
python-engineio==3.13.2
python-socketio==4.6.1
python-telegram-bot==12.8
pytz==2020.5
PyYAML==6.0.1
questionary==1.5.2
rasa==2.1.2
rasa-sdk==2.3.1
redis==3.5.3
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rfc3986==1.5.0
rocketchat-API==1.9.1
rsa==4.9
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.8
s3transfer==0.8.2
safetensors==0.4.1
sanic==20.9.0
Sanic-Cors==0.10.0.post3
sanic-jwt==1.4.1
Sanic-Plugins-Framework==0.9.5
scikit-learn==0.23.2
scipy==1.7.3
sentence-transformers==2.2.2
sentencepiece==0.1.99
sentry-sdk==0.19.5
six==1.16.0
sklearn-crfsuite==0.3.6
slackclient==2.9.4
sniffio==1.3.0
SQLAlchemy==1.3.24
tabulate==0.9.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.3.4
tensorflow-addons==0.12.0
tensorflow-estimator==2.3.0
tensorflow-hub==0.9.0
tensorflow-probability==0.11.1
termcolor==2.3.0
terminaltables==3.1.10
threadpoolctl==3.1.0
tokenizers==0.13.3
torch==1.13.1
torchvision==0.14.1
tornado==6.2
tqdm==4.50.2
transformers==4.30.2
twilio==6.45.4
typeguard==2.13.3
typing-extensions==3.10.0.2
tzdata==2023.3
tzlocal==5.1
ujson==3.2.0
urllib3==1.26.18
wcwidth==0.2.12
webexteamssdk==1.6.1
websockets==8.1
Werkzeug==2.2.3
wincertstore==0.2
wrapt==1.16.0
yarl==1.5.1
zipp==3.15.0
